Extract shutdown related functions to utils so they can be used by launcher
