+++ pkg/virt-handler/vm_test.go
@@ -312,78 +312,6 @@ var _ = Describe("VirtualMachineInstance", func() {
 			testutils.ExpectEvent(recorder, VMIStopping)
 		})
 
-		It("should handle cleanup of legacy graceful shutdown and watchdog files", func() {
-
-			name := "testvmi"
-			namespace := "default"
-			uid := "1234"
-
-			legacyMockSockFile := filepath.Join(shareDir, "sockets", uid+"_sock")
-			clientInfo := &virtcache.LauncherClientInfo{
-				Client:             client,
-				SocketFile:         legacyMockSockFile,
-				DomainPipeStopChan: make(chan struct{}),
-				Ready:              true,
-			}
-			controller.addLauncherClient(types.UID(uid), clientInfo)
-			err := virtcache.AddGhostRecord(namespace, name, legacyMockSockFile, types.UID(uid))
-			Expect(err).ToNot(HaveOccurred())
-
-			os.MkdirAll(filepath.Dir(legacyMockSockFile), 0755)
-			os.MkdirAll(filepath.Join(shareDir, "watchdog-files"), 0755)
-			os.MkdirAll(filepath.Join(shareDir, "graceful-shutdown-trigger"), 0755)
-
-			watchdogFile := filepath.Join(shareDir, "watchdog-files", namespace+"_"+name)
-			gracefulTrigger := filepath.Join(shareDir, "graceful-shutdown-trigger", namespace+"_"+name)
-
-			f, err := os.Create(legacyMockSockFile)
-			Expect(err).ToNot(HaveOccurred())
-			f.Close()
-
-			f, err = os.Create(watchdogFile)
-			Expect(err).ToNot(HaveOccurred())
-			_, err = f.WriteString(uid)
-			Expect(err).ToNot(HaveOccurred())
-			f.Close()
-
-			f, err = os.Create(gracefulTrigger)
-			Expect(err).ToNot(HaveOccurred())
-			f.Close()
-
-			exists, err := diskutils.FileExists(filepath.Join(shareDir, "sockets"))
-			Expect(err).ToNot(HaveOccurred())
-			Expect(exists).To(BeTrue())
-
-			mockQueue.Add(namespace + "/" + name)
-			mockHotplugVolumeMounter.EXPECT().UnmountAll(gomock.Any()).Return(nil)
-			client.EXPECT().Close()
-			controller.Execute()
-
-			exists, err = diskutils.FileExists(filepath.Join(shareDir, "sockets"))
-			Expect(err).ToNot(HaveOccurred())
-			Expect(exists).To(BeTrue())
-
-			exists, err = diskutils.FileExists(filepath.Join(shareDir, "graceful-shutdown-trigger"))
-			Expect(err).ToNot(HaveOccurred())
-			Expect(exists).To(BeTrue())
-
-			exists, err = diskutils.FileExists(filepath.Join(shareDir, "watchdog-files"))
-			Expect(err).ToNot(HaveOccurred())
-			Expect(exists).To(BeTrue())
-
-			exists, err = diskutils.FileExists(legacyMockSockFile)
-			Expect(err).ToNot(HaveOccurred())
-			Expect(exists).To(BeFalse())
-
-			exists, err = diskutils.FileExists(watchdogFile)
-			Expect(err).ToNot(HaveOccurred())
-			Expect(exists).To(BeFalse())
-
-			exists, err = diskutils.FileExists(gracefulTrigger)
-			Expect(err).ToNot(HaveOccurred())
-			Expect(exists).To(BeFalse())
-		})
-
 		It("should attempt graceful shutdown of Domain if no cluster wide equivalent exists", func() {
 			vmi := api2.NewMinimalVMI("testvmi")
 			vmi.UID = vmiTestUUID
